+++ .env.example
@@ -3,3 +3,14 @@ SP1_PROVER=mock
 # If using the proving network, set to your whitelisted private key. For more information, see:
 # https://docs.succinct.xyz/prover-network/setup.html#key-setup
 SP1_PRIVATE_KEY=
+
+# L1 specific (can be eth/sepolia if proving for pos mainnet or amoy)
+CHAIN_ID= # chain id
+ETH_RPC_URL= # eth rpc url
+PRIVATE_KEY= # private key to send proofs
+CONTRACT_ADDRESS= # verifier contract address
+
+# pos specific
+HEIMDALL_REST_ENDPOINT="http://localhost:1317" # heimdall rest endpoint
+TENDERMINT_ENDPOINT="http://localhost:25567" # tendermint endpoint
+BOR_RPC_URL="http://localhost:8545" # bor rpc url
