+++ .gitignore
@@ -20,3 +20,7 @@ pgo-data.profdata
 
 # Succinct programs
 program/elf/riscv32im-succinct-zkvm-elf
+elf/riscv32im-succinct-zkvm-elf
+
+# Added by cargo
+/target
