+++ .vscode/settings.json
@@ -1,7 +1,6 @@
 {
     "rust-analyzer.linkedProjects": [
         "program/Cargo.toml",
-        "script/Cargo.toml",
         "consensus-proof/Cargo.toml",
         "operator/Cargo.toml",
     ],
