+++ Cargo.toml
@@ -1,0 +1,138 @@
+[workspace]
+members = [
+    "consensus-proof",
+    "operator",
+    "program",
+]
+resolver = "2"
+
+[workspace.dependencies]
+base64 = "0.22.1"
+sha2 = "0.10.8"
+bytes = "1.7.1"
+prost = "0.13.1"
+prost-types = "0.13.1"
+bincode = "1.3.3"
+anyhow = "1.0.82"
+eyre = "0.6"
+ethers = "2.0.14"
+dotenv = "0.15.0"
+clap = "4.5.18"
+url = "2.3"
+serde_json = { version = "1.0", default-features = false, features = ["alloc"] }
+serde = { version = "1.0", default-features = false, features = ["derive"] }
+reqwest = { version = "0.12.7", features = ["json"] }
+tokio = { version = "1.21", default-features = false, features = [
+    "rt",
+    "rt-multi-thread",
+] }
+
+# reth
+reth-primitives = { git = "https://github.com/sp1-patches/reth", tag = "rsp-20240830", default-features = false, features = [
+    "alloy-compat",
+    "std",
+    "secp256k1"
+] }
+
+# alloy
+alloy-primitives = "0.8.0"
+alloy-rlp = "0.3.4"
+alloy-sol-types = { version = "0.8" }
+alloy-provider = { version = "0.3", default-features = false, features = [
+    "reqwest",
+] }
+alloy-sol-macro = { version = "0.8" }
+alloy = { version = "0.3" }
+alloy-rpc-types = { version = "0.3", default-features = false, features = [
+    "eth",
+] }
+
+# sp1
+sp1-zkvm = "2.0.0"
+sp1-sdk = { git = "https://github.com/succinctlabs/sp1.git", tag = "v2.0.0" }
+sp1-cc-client-executor = { git = "https://github.com/succinctlabs/sp1-contract-call/", branch = "main" }
+sp1-cc-host-executor = { git = "https://github.com/succinctlabs/sp1-contract-call/", branch = "main" }
+
+# [build-dependencies]
+# prost-build = "0.13.1"
+# sp1-helper = { git = "https://github.com/succinctlabs/sp1.git", tag = "v2.0.0" }
+
+[patch.crates-io]
+crypto-bigint = { git = "https://github.com/sp1-patches/RustCrypto-bigint", branch = "patch-v0.5.5" }
+tiny-keccak = { git = "https://github.com/sp1-patches/tiny-keccak", branch = "patch-v2.0.2" }
+ecdsa-core = { git = "https://github.com/sp1-patches/signatures", package = "ecdsa", branch = "patch-ecdsa-v0.16.9" }
+secp256k1 = { git = "https://github.com/sp1-patches/rust-secp256k1", branch = "patch-secp256k1-v0.29.1" }
+substrate-bn = { git = "https://github.com/sp1-patches/bn", branch = "patch-v0.6.0" }
+
+# [workspace.dependencies]
+# serde_json = "1.0.94"
+# serde = { version = "1.0", default-features = false, features = ["derive"] }
+# url = "2.3"
+# clap = { version = "4.5.7", features = ["derive", "env"] }
+# dotenv = "0.15.0"
+# bincode = "1.3.3"
+# eyre = "0.6"
+# tokio = { version = "1.21", default-features = false, features = [
+#     "rt",
+#     "rt-multi-thread",
+# ] }
+# hex = "0.4.3"
+
+
+# # rsp
+# rsp-rpc-db = { git = "https://github.com/invocamanman/rsp" , branch = "feature/addDinamycChainIDClique"}
+# rsp-witness-db = { git = "https://github.com/invocamanman/rsp" , branch = "feature/addDinamycChainIDClique"}
+# rsp-primitives = { git = "https://github.com/invocamanman/rsp" , branch = "feature/addDinamycChainIDClique"}
+# rsp-client-executor = { git = "https://github.com/invocamanman/rsp" , branch = "feature/addDinamycChainIDClique"}
+# rsp-mpt = { git = "https://github.com/invocamanman/rsp" , branch = "feature/addDinamycChainIDClique"}
+
+# rsp-host-executor = { git = "https://github.com/invocamanman/rsp", branch = "feature/addDinamycChainIDClique" }
+
+# # sp1
+# sp1-sdk = "2.0.0"
+# sp1-zkvm = {version = "2.0.0", features = ["verify"]}
+
+
+# # alloy
+# alloy-provider = { version = "0.3", default-features = false, features = [
+#     "reqwest",
+#     "reqwest-rustls-tls",
+# ] }
+# alloy-sol-types = { version = "0.8" }
+# alloy-primitives = "0.8.0"
+# alloy-sol-macro = { version = "0.8" }
+# alloy = { version = "0.3" }
+# alloy-rpc-types = { version = "0.3", default-features = false, features = [
+#     "eth",
+# ] }
+
+# # revm
+# revm = { version = "=14.0.1", features = [
+#     "optimism",
+#     "std",
+#     "serde",
+#     "kzg-rs",
+# ], default-features = false }
+
+# # reth
+# reth-primitives = { git = "https://github.com/sp1-patches/reth", tag = "rsp-20240830", default-features = false, features = [
+#     "alloy-compat",
+#     "optimism",
+#     "std",
+# ] }
+
+# # sp1-cc
+# sp1-cc-client-executor = { git = "https://github.com/succinctlabs/sp1-contract-call/", branch = "main" }
+# sp1-cc-host-executor = { git = "https://github.com/succinctlabs/sp1-contract-call/", branch = "main" }
+
+
+# # sha2
+# sha2 = "0.10.8"
+
+# [patch.crates-io]
+# sha2-v0-10-8 = { git = "https://github.com/sp1-patches/RustCrypto-hashes", package = "sha2", branch = "patch-sha2-v0.10.8" }
+# sha3-v0-10-8 = { git = "https://github.com/sp1-patches/RustCrypto-hashes", package = "sha3", branch = "patch-sha3-v0.10.8" }
+# crypto-bigint = { git = "https://github.com/sp1-patches/RustCrypto-bigint", branch = "patch-v0.5.5" }
+# tiny-keccak = { git = "https://github.com/sp1-patches/tiny-keccak", branch = "patch-v2.0.2" }
+# ecdsa-core = { git = "https://github.com/sp1-patches/signatures", package = "ecdsa", branch = "patch-ecdsa-v0.16.9" }
+# substrate-bn = { git = "https://github.com/sp1-patches/bn", branch = "patch-v0.6.0" }
