+++ consensus-proof/Cargo.toml
@@ -1,36 +1,24 @@
-[workspace]
 [package]
 version = "0.1.0"
 name = "pos-consensus-proof"
 edition = "2021"
 
 [dependencies]
-base64 = "0.22.1"
-sha2 = "0.10.8"
-bytes = "1.7.1"
-prost = "0.13.1"
-prost-types = "0.13.1"
-bincode = "1.3.3"
+base64.workspace = true
+sha2.workspace = true
+bytes.workspace = true
+prost.workspace = true
+prost-types.workspace = true
+bincode.workspace = true
 
 # reth
-reth-primitives = { git = "https://github.com/sp1-patches/reth", tag = "rsp-20240830", default-features = false, features = [
-    "alloy-compat",
-    "std",
-    "secp256k1"
-] }
+reth-primitives.workspace = true
 
 # alloy
-alloy-primitives = "0.8.0"
-alloy-rlp = "0.3.4"
-alloy-sol-types = { version = "0.8" }
-sp1-cc-client-executor = { git = "https://github.com/succinctlabs/sp1-contract-call/", branch = "main" }
+alloy-primitives.workspace = true
+alloy-rlp.workspace = true
+alloy-sol-types.workspace = true
+sp1-cc-client-executor.workspace = true
 
 [build-dependencies]
 prost-build = "0.13.1"
-
-[patch.crates-io]
-crypto-bigint = { git = "https://github.com/sp1-patches/RustCrypto-bigint", branch = "patch-v0.5.5" }
-tiny-keccak = { git = "https://github.com/sp1-patches/tiny-keccak", branch = "patch-v2.0.2" }
-ecdsa-core = { git = "https://github.com/sp1-patches/signatures", package = "ecdsa", branch = "patch-ecdsa-v0.16.9" }
-secp256k1 = { git = "https://github.com/sp1-patches/rust-secp256k1", branch = "patch-secp256k1-v0.29.1" }
-substrate-bn = { git = "https://github.com/sp1-patches/bn", branch = "patch-v0.6.0" }
