+++ consensus-proof/src/types.rs
@@ -86,8 +86,8 @@ mod tests {
     use reth_primitives::hex;
     use std::str::FromStr;
 
-    // #[test]
-    fn test_1() {
+    #[test]
+    fn test_deserialize_msg() {
         let a = "6gHwYl3uCqAB0ss+ZgoUCSB6bv7jRss+SlSsGFI+NxXTiz8Q/KaMBhj6p4wGIiBhk6zRTSThGAsyswmIseJyY9Eg8rrnHi4vXGNFGJ/r9SoFODAwMDIyUTUwMjM0MTM1LWQ5YmUtNGU0YS04NGY3LTM1OTZjZmIwN2EwZCAtIDB4ODhiMWUyNzI2M2QxMjBmMmJhZTcxZTJlMmY1YzYzNDUxODlmZWJmNRJB5jp3Zv4MQiiaOQ612UlPgyJzjt3v5YAJs9sqArSSsXVnssdRf5as1uuwettRNPGFlohE8saPapGLQxF74mHm/AE=".to_string();
         let mut decoded_tx_data = BASE64_STANDARD.decode(a).expect("tx_data decoding failed");
         let decoded_message = deserialize_msg(&mut decoded_tx_data).unwrap();
@@ -112,53 +112,8 @@ mod tests {
             signature: sig,
             memo: "".to_string(),
         };
-
-        let s = serialize_msg(&msg);
-        println!("serialized: {:?}", s);
-
         assert_eq!(decoded_message, msg);
     }
-
-    #[test]
-    fn test_2() {
-        println!("test2...");
-        let a = "6gHwYl3uCqAB0ss+ZgoUbcLdVPJJeewmISeUxxr+/tciKAwQic2OBhiWzY4GIiCRtK1d5icJ6mniXvKka6jQ3kSRzEqtjzXvFLTdIWYdVCoFODAwMDIyUTU1MjA1NmNlLWVlYzktNGQ3Mi1iOWVhLTE4Y2MzNDNlY2U2MSAtIDB4YTQ2YmE4ZDBkZTQ0OTFjYzRhYWQ4ZjM1ZWYxNGI0ZGQyMTY2MWQ1NBJBc53ZW0oTiwuf5z7XcbuSDqqsoHhEf/vwcsNpk89Az7QMnTwpLfZLexmIbCN8BA6CgxXActL96FViORpHrnJ2FQA=".to_string();
-        let mut decoded_tx_data = BASE64_STANDARD.decode(a).expect("tx_data decoding failed");
-        let decoded_message = deserialize_msg(&mut decoded_tx_data).unwrap();
-        println!("{:?}", decoded_message);
-    }
-
-    // #[test]
-    // fn test_deserialize_msg() {
-    //     println!("hiiii???");
-    //     let decoded_str = "0xe801f0625dee0a9e01d2cb3e660a14fcccd43296d9c1601a904eca9b339d94a5e5e09810f8b0841d188ab1841d22207520ee2c289b7ecf623d4f8a44dc6ad772d92ee4375a2014dabea80e7ef8d55222a03313337325164386430396366342d663735662d343864332d386565372d663263323130636237323733202d2030783434646336616437373264393265653433373561323031346461626561383065376566386435353212417bc767635eb060d2fc42ad3aa67cd0f1991ef1412fc9c28abc1c4eac4700b11d153d6b3258fe29b8e8674a36afdcc5c0203e01987f062fa9fe1ce950265bed2f00";
-    //     let mut decoded_bytes = hex::decode(decoded_str).unwrap();
-
-    //     let decoded_msg = deserialize_msg(&mut decoded_bytes).unwrap();
-
-    //     let m = heimdall_types::std_tx::Msg::Milestone(heimdall_types::MilestoneMsg {
-    //         proposer: hex::decode("FCCCD43296D9C1601A904ECA9B339D94A5E5E098")
-    //             .unwrap()
-    //             .to_vec(),
-    //         start_block: 60889208,
-    //         end_block: 60889226,
-    //         hash: hex::decode("7520EE2C289B7ECF623D4F8A44DC6AD772D92EE4375A2014DABEA80E7EF8D552")
-    //             .unwrap()
-    //             .to_vec(),
-    //         bor_chain_id: "137".to_string(),
-    //         milestone_id:
-    //             "d8d09cf4-f75f-48d3-8ee7-f2c210cb7273 - 0x44dc6ad772d92ee4375a2014dabea80e7ef8d552"
-    //                 .to_string(),
-    //     });
-    //     let sig = hex::decode("0x7bc767635eb060d2fc42ad3aa67cd0f1991ef1412fc9c28abc1c4eac4700b11d153d6b3258fe29b8e8674a36afdcc5c0203e01987f062fa9fe1ce950265bed2f00").unwrap().to_vec();
-    //     let msg = heimdall_types::StdTx {
-    //         msg: Some(m),
-    //         signature: sig,
-    //         memo: "".to_string(),
-    //     };
-
-    //     assert_eq!(decoded_msg, msg);
-    // }
 
     #[test]
     fn test_precommit_msg() {
@@ -199,7 +154,7 @@ mod tests {
         assert_eq!(decoded, vote);
     }
 
-    // #[test]
+    #[test]
     fn test_validator_set() {
         let hex_msg = "0a600801200128904e324104dc19fdf9a82fd5c4327f31b96b6bbe0b9d44564ad89c2139db47c5cb2def87ac584fc05117663de2f17ae5ee50eced7283a596e10aaf33fb34c4cf5f98e4fda73a146ab3d36c46ecfb9b9c0bd51cb1c3da5a2c81cea612600801200128904e324104dc19fdf9a82fd5c4327f31b96b6bbe0b9d44564ad89c2139db47c5cb2def87ac584fc05117663de2f17ae5ee50eced7283a596e10aaf33fb34c4cf5f98e4fda73a146ab3d36c46ecfb9b9c0bd51cb1c3da5a2c81cea6";
         let mut bytes_msg = hex::decode(hex_msg).unwrap();
