+++ operator/Cargo.toml
@@ -4,60 +4,46 @@ version = "0.1.0"
 edition = "2021"
 
 [[bin]]
-name = "repro"
-path = "src/bin/repro.rs"
+name = "vkey"
+path = "src/bin/vkey.rs"
 
-# [[bin]]
-# name = "vkey"
-# path = "src/bin/vkey.rs"
+[[bin]]
+name = "verify"
+path = "src/bin/verify.rs"
 
 [dependencies]
-serde_json = { version = "1.0", default-features = false, features = ["alloc"] }
-serde = { version = "1.0", default-features = false, features = ["derive"] }
-# tokio = { version = "1.40.0", features = ["full"] } 
-tokio = { version = "1.21", default-features = false, features = [
-    "rt",
-    "rt-multi-thread",
-] }
-reqwest = { version = "0.12.7", features = ["json"] }
-anyhow = "1.0.82"
-eyre = "0.6"
-ethers = "2.0.14"
-dotenv = "0.15.0"
-clap = "4.5.18"
-bincode = "1.3.3"
-prost-types = "0.13.1"
-base64 = "0.22.1"
-url = "2.3"
 pos-consensus-proof = { path = "../consensus-proof" }
 
+serde_json.workspace = true
+serde.workspace = true
+tokio.workspace = true
+reqwest.workspace = true
+anyhow.workspace = true
+eyre.workspace = true
+ethers.workspace = true
+dotenv.workspace = true
+clap.workspace = true
+bincode.workspace = true
+prost-types.workspace = true
+base64.workspace = true
+url.workspace = true
+
 # sp1
-sp1-sdk = { git = "https://github.com/succinctlabs/sp1.git", tag = "v2.0.0" }
-sp1-cc-client-executor = { git = "https://github.com/succinctlabs/sp1-contract-call/", branch = "main" }
-sp1-cc-host-executor = { git = "https://github.com/succinctlabs/sp1-contract-call/", branch = "main" }
+sp1-sdk.workspace = true
+sp1-cc-client-executor.workspace = true
+sp1-cc-host-executor.workspace = true
 
 # reth
-reth-primitives = { git = "https://github.com/sp1-patches/reth", tag = "rsp-20240830", default-features = false, features = [
-    "alloy-compat",
-    "std",
-    "secp256k1"
-] }
+reth-primitives.workspace = true
 
 # alloy
-alloy-provider = { version = "0.3", default-features = false, features = [
-    "reqwest",
-] }
-alloy-primitives = "0.8.0"
-alloy-sol-types = { version = "0.8" }
-alloy-sol-macro = { version = "0.8" }
-alloy = { version = "0.3" }
-alloy-rpc-types = { version = "0.3", default-features = false, features = [
-    "eth",
-] }
+alloy-provider.workspace = true
+alloy-primitives.workspace = true
+alloy-sol-types.workspace = true
+alloy-sol-macro.workspace = true
+alloy.workspace = true
+alloy-rpc-types.workspace = true
 
 [build-dependencies]
-sp1-helper = { git = "https://github.com/succinctlabs/sp1.git", tag = "v2.0.0" }
 prost-build = "0.13.1"
-
-# [patch.crates-io]
-# alloy-eips = { git = "https://github.com/jtguibas/alloy.git", branch = "john/rsp-8e9e6ac" }
+sp1-helper = { git = "https://github.com/succinctlabs/sp1.git", tag = "v2.0.0" }
