--- operator/src/bin/main.rs
@@ -1,14 +1,0 @@
-use sp1_sdk::{HashableKey, ProverClient};
-
-/// The ELF (executable and linkable format) file for the Succinct RISC-V zkVM.
-///
-/// This file is generated by running `cargo prove build` inside the `program` directory.
-pub const CONSENSUS_PROOF_ELF: &[u8] =
-    include_bytes!("../../../program/elf/riscv32im-succinct-zkvm-elf");
-
-// #[tokio::main]
-fn main() {
-    let client = ProverClient::new();
-    let (_pk, vk) = client.setup(CONSENSUS_PROOF_ELF);
-    print!("vk: {:?}", vk.bytes32());
-}
