+++ operator/src/bin/vkey.rs
@@ -2,7 +2,7 @@
 //!
 //! You can run this script using the following command:
 //! ```shell
-//! RUST_LOG=info cargo run --package fibonacci-script --bin vkey --release
+//! RUST_LOG=info cargo run --package operator --bin vkey --release
 //! ```
 
 use sp1_sdk::{HashableKey, ProverClient};
@@ -10,7 +10,8 @@ use sp1_sdk::{HashableKey, ProverClient};
 /// The ELF (executable and linkable format) file for the Succinct RISC-V zkVM.
 ///
 /// This file is generated by running `cargo prove build` inside the `program` directory.
-pub const FIBONACCI_ELF: &[u8] = include_bytes!("../../../program/elf/riscv32im-succinct-zkvm-elf");
+pub const CONSENSUS_PROOF_ELF: &[u8] =
+    include_bytes!("../../../program/elf/riscv32im-succinct-zkvm-elf");
 
 fn main() {
     // Setup the logger.
@@ -20,7 +21,7 @@ fn main() {
     let client = ProverClient::new();
 
     // Setup the program.
-    let (_, vk) = client.setup(FIBONACCI_ELF);
+    let (_, vk) = client.setup(CONSENSUS_PROOF_ELF);
 
     // Print the verification key.
     println!("Program Verification Key: {}", vk.bytes32());
