+++ program/Cargo.toml
@@ -1,23 +1,12 @@
-[workspace]
 [package]
 version = "0.1.0"
 name = "consensus-proof-program"
 edition = "2021"
 
 [dependencies]
-sp1-zkvm = "2.0.0"
 pos-consensus-proof = { path = "../consensus-proof" }
-reth-primitives = { git = "https://github.com/sp1-patches/reth", tag = "rsp-20240830", default-features = false, features = [
-    "alloy-compat",
-    "std",
-    "secp256k1"
-] }
-alloy-primitives = "0.8.0"
-alloy-sol-types = { version = "0.8" }
 
-[patch.crates-io]
-crypto-bigint = { git = "https://github.com/sp1-patches/RustCrypto-bigint", branch = "patch-v0.5.5" }
-tiny-keccak = { git = "https://github.com/sp1-patches/tiny-keccak", branch = "patch-v2.0.2" }
-ecdsa-core = { git = "https://github.com/sp1-patches/signatures", package = "ecdsa", branch = "patch-ecdsa-v0.16.9" }
-secp256k1 = { git = "https://github.com/sp1-patches/rust-secp256k1", branch = "patch-secp256k1-v0.29.1" }
-substrate-bn = { git = "https://github.com/sp1-patches/bn", branch = "patch-v0.6.0" }
+sp1-zkvm.workspace = true
+reth-primitives.workspace = true
+alloy-primitives.workspace = true
+alloy-sol-types.workspace = true
