--- script/Cargo.toml
@@ -1,24 +1,0 @@
-[workspace]
-[package]
-version = "0.1.0"
-name = "consensus-proof-script"
-edition = "2021"
-default-run = "prove"
-
-[[bin]]
-name = "prove"
-path = "src/bin/prove.rs"
-
-[[bin]]
-name = "vkey"
-path = "src/bin/vkey.rs"
-
-[dependencies]
-sp1-sdk = { git = "https://github.com/succinctlabs/sp1.git", tag = "v2.0.0" }
-pos-consensus-proof = { path = "../consensus-proof" }
-
-[build-dependencies]
-sp1-helper = { git = "https://github.com/succinctlabs/sp1.git", tag = "v2.0.0" }
-
-[patch.crates-io]
-alloy-eips = { git = "https://github.com/jtguibas/alloy.git", branch = "john/rsp-8e9e6ac" }
